Refactor: extract type visitors and deduplicate operator stringification
Move TypeVariableUpdater and FreeTypeVariableCollector from
type_environment.hpp into new type_visitors.hpp. Remove duplicate
operator-to-string methods from dump.hpp in favor of stream operators
from operators.hpp.
