+++ bust/test/src/type_unifier_test.cpp
@@ -13,8 +13,8 @@
 //*
 //****************************************************************************
 
-#include <hir/type_environment.hpp>
 #include <hir/type_unifier.hpp>
+#include <hir/type_visitors.hpp>
 #include <hir/types.hpp>
 
 #include <doctest/doctest.h>
